# Fix inconsistencies of findnext and findprev functions