… base/strings/util.jl
… test/strings/search.jl
… test/strings/util.jl