This pull request fixes several inconsistent behaviors . The proposed changes are breaking in some edge cases.

There are mainly three behavioral changes:
- Consistent handling of empty strings (reported in #39940 and #36768).
- More permissive bounds checking.
- Automatic index alignment.

## Empty strings

The first change will be best depicted by the following example.

master:
```
julia> [findnext("", "abc", i) for i in 1:4]  # This is fine.
4-element Vector{UnitRange{Int64}}:
 1:0
 2:1
 3:2
 4:3

julia> [findprev("", "abc", i) for i in 0:3]  # Why duplicated 1:0 ?
4-element Vector{UnitRange{Int64}}:
 1:0
 1:0
 2:1
 3:2
```

proposed:
```
julia> [findnext("", "abc", i) for i in 1:4]  # no changes
4-element Vector{UnitRange{Int64}}:
 1:0
 2:1
 3:2
 4:3

julia> [findprev("", "abc", i) for i in 0:3]  # now consistent
4-element Vector{UnitRange{Int64}}:
 1:0
 2:1
 3:2
 4:3
```

This also changes `findlast("", b)` if `b` is not an empty string.
```
julia> findlast("", "abc")  # master
3:2

julia> findlast("", "abc")  # proposed
4:3
```

## Bounds check

The second change makes it possible to specify any position larger than the last index of the second string:

master:
```
julia> findnext("a", "abc", 4)

julia> findnext("a", "abc", 5)
ERROR: BoundsError: attempt to access 3-codeunit String at index [5]

julia> findnext("a", "abc", 6)
ERROR: BoundsError: attempt to access 3-codeunit String at index [6]
```

proposed:
```
julia> findnext("a", "abc", 4)

julia> findnext("a", "abc", 5)

julia> findnext("a", "abc", 6)
```

I'm not sure why the current behavior is so strict, but the proposed behavior is more consistent with the following generic method of `findnext`:
https://github.com/JuliaLang/julia/blob/6913f9ccb156230f54830c8b7b8ef82b41cac27e/base/array.jl#L1855-L1866

`findprev` is also changed in the symmetric way.

master:
```
julia> findprev("a", "abc", 0)

julia> findprev("a", "abc", -1)
ERROR: BoundsError: attempt to access 3-codeunit String at index [-1]

julia> findprev("a", "abc", -2)
ERROR: BoundsError: attempt to access 3-codeunit String at index [-2]
```

proposed:
```
julia> findprev("a", "abc", 0)

julia> findprev("a", "abc", -1)

julia> findprev("a", "abc", -2)
```

## Index alignment

The third change allows staring a search in the middle of a character in multibyte strings.

master:
```
julia> findnext("β", "αβ", 2)
ERROR: StringIndexError: invalid index [2], valid nearby indices [1]=>'α', [3]=>'β'
```

proposed:
```
julia> findnext("β", "αβ", 2)
3:3
```

This change simplify some code that uses `findnext` or `findprev`. For example, `findall` can be implemented in an obvious way:
```julia
function findall(a, b)
    result = UnitRange{Int}[]
    r = findfirst(a, b)
    while r !== nothing
        push!(result, r)
        r = findnext(a, b, first(r) + 1)  # no need to align the index with `nextind` 
    end
    return result
end
```

In my opinion, the index check is a kind of implementation detail and should be hidden from the user.

## Performance

I also refactored the current code almost from scratch because it was difficult to fix inconsistencies without that. So, I quickly checked the performance degradation and discovered that the performance was slightly improved for string-string search.

This is a benchmark script.
```julia
using BenchmarkTools

b = "abracadabra"
for a in ['c', "cad"]
    @show a
    @btime findfirst($a, $b)
    @btime findlast($a, $b)
end
println()

b = "Julia is a high-level, high-performance dynamic language for technical computing.  The main homepage for Julia can be found at https://julialang.org/.  This is the GitHub repository of Julia source code, including instructions for compiling and installing Julia, below."
for a in ['H', "instructions"]
    @show a
    @btime findfirst($a, $b)
    @btime findlast($a, $b)
end
```

master:
```
a = 'c'
  9.456 ns (0 allocations: 0 bytes)
  10.872 ns (0 allocations: 0 bytes)
a = "cad"
  27.580 ns (2 allocations: 128 bytes)
  38.288 ns (2 allocations: 128 bytes)

a = 'H'
  11.613 ns (0 allocations: 0 bytes)
  10.871 ns (0 allocations: 0 bytes)
a = "instructions"
  89.698 ns (2 allocations: 128 bytes)
  55.495 ns (2 allocations: 128 bytes)
```

proposed:
```
a = 'c'
  7.572 ns (0 allocations: 0 bytes)
  10.169 ns (0 allocations: 0 bytes)
a = "cad"
  17.066 ns (0 allocations: 0 bytes)
  16.152 ns (0 allocations: 0 bytes)

a = 'H'
  9.457 ns (0 allocations: 0 bytes)
  10.489 ns (0 allocations: 0 bytes)
a = "instructions"
  74.781 ns (0 allocations: 0 bytes)
  28.898 ns (0 allocations: 0 bytes)
```

## Other problems

I found several questionable behaviors while working on this pull request; some of them seem to be intentional because they are explicitly tested. One is about splitting a string with an empty string, which is already filed at #40117. Another is the following two test cases:
https://github.com/JuliaLang/julia/blob/6913f9ccb156230f54830c8b7b8ef82b41cac27e/test/strings/search.jl#L420-L421

This suggests `findprev` should be more permissive in the right bound than `findnext` should be in the left bound (`findnext("a", "abc", 0)` throws `BoundsError`). This is also expected in other places such as https://github.com/JuliaLang/julia/blob/6913f9ccb156230f54830c8b7b8ef82b41cac27e/test/strings/search.jl#L319. This seems intended but is not consistent. So, I've changed the behavior in the current proposal so that `findprev("a", "abc", 4)` (for example) throws an `BoundsError` exception, which is consistent with `findnext` and other methods of the `findprev` function. Technically speaking, this is a breaking change, and thus I'm not perfectly sure if this is acceptable.