@oscardssmith 2021-03-21
This might be kind of dumb, but could we do the string versions of this by reinterpreting to `Array{UInt8}` and using those methods? Since the indexing is the same, I think it might just work (and let us reduce a bunch of code).