@bicycle1885 2021-03-21
> could we do the string versions of this by reinterpreting to Array{UInt8} and using those methods?

In fact, the original code does that by wrapping a string `s` with `unsafe_wrap(Vector{UInt8}, s)`. The problem is that is allocates an object for each string and the cost (~5 ns) is not negligible especially for short strings. I guess the performance improvement of this pull request is largely thanks to avoiding object allocations. Also, considering that the amount of code has been reduced by the proposed change, the effect of code reduction seems limited.