@JeffBezanson 2021-03-25
`findprev("", "abc", 1) === 2:1` doesn't entirely make sense to me (and some others) --- how does going "previous" give a higher index?