@bicycle1885 2021-03-25
This is the documented behavior (see the last line):
https://github.com/JuliaLang/julia/blob/1897e08ff54eb8c91b806e20bbfecbe7b2d4f9f7/base/strings/search.jl#L534-L542

That is, the third argument of `findprev` specifies the upper bound of the last index of a returned range. Of course, we can break this constraint for empty strings, but isn't it inconsistent?