@bicycle1885 2021-03-26
I think `r = findprev(a, b, i)` should return the rightmost range that satisfies `a == b[r] && last(r) <= i` (this constraint is documented). If we stick to this rule, `findprev("", "abc", 1) === 2:1` is the only consistent behavior.

https://github.com/JuliaLang/julia/blob/1897e08ff54eb8c91b806e20bbfecbe7b2d4f9f7/base/strings/search.jl#L534-L542