@Seelengrab 2022-07-21
The examples, fixes & consistency improvements look good, so I'd be in favor, but I haven't looked at the code in detail. Would need a rebase & some squashing though.

@oscardssmith do you happen to remember what there was to AAAAAAARGH about?