@oscardssmith 2022-07-21
basically the problem is that there doesn't seem to be a good api that is fully consistent.