@Moelf 2023-05-25
>I found several questionable behaviors while working on this pull request; some of them seem to be intentional because they are explicitly tested. ... Another is the following two test cases:

yeah, we added those tests here:
- https://github.com/JuliaLang/julia/pull/37283

and I remember we added those tests because some code rely on that behavior and earlier version of that PR broke those code, thus we added a test.