@oscardssmith 2023-05-25
> findprev("", "abc", 1) === 2:1 doesn't entirely make sense to me (and some others) --- how does going "previous" give a higher index?

Triage was re-looking at this and the answer here is that for `findfirst` the `start` argument provides a lower bound on the first index, while symetrically, `findprev` should have the invariant that the `start` argument (possibly should be renamed) provides an upper bound on the last index.