@LilithHafner 2023-05-25
IIUC, in all cases,
`findprev(predicate, collection, start)` should return the last valid index less than or equal to start for which the predicate returns true.
`findnext(predicate, collection, start)` should return the first valid index greater than or equal to start for which the predicate returns true.