@Moelf 2023-05-26
>We also should make findnext and findprev for Vector have the same semantics as currently Vector{UInt8} behaves like String while all other Vectors don't let you go out of bounds at all.

I wonder if people are really comfortable with `Vector{UInt8}` being different than other `Vector{X}`, maybe the solution is to have this behavior for `Base.CodeUnits{UInt8, <:AbstractString}`? (which means we will deprecate what we added in #37283)