@oscardssmith 2023-05-26
this proposal is to make the semantics of all vectors consistent with the semantics for strings.