@oscardssmith 2023-05-26
yeah. sorry if that was unclear. I think @LilithHafner's description of triage intent is exactly correct. for all collection types, if there is a matching value before the index, findprev will return it, and will otherwise return nothing. (and likewise for findnext/after)